requirements.txt
streamlit-elements
markdown-analysis
openai
google-generativeai